A Ruby program that will return the equivalent rate of Philippine peso to a provided country.
Currency converter is a Ruby program that will accept a country name as
an argument. The program will access a soap web service by using the
soap4r gem. To get the currency code of the supplied country, the
program will create a hash from countries.xml that contains data fetched
from http://www.geonames.org/ and generated by
http://peric.github.io/GetCountries/. To get the answer, a 'parameters'
hash is passed as parameters in calling the driver method,
conversionRate and the conversionRateResult method.

Please use ruby version <= 1.8.7. This is because the 'soap/wsdlDriver'
library was removed in the newer versions.